Rename confluent_kafka_shared to common
